added users and chats resources authorization

+++ data/api/resources/chat_resource.py
@@ -4,17 +4,20 @@ from flask_restful import Resource, abort
 from ... import db_session
 from ...models.chats import Chat
 
-from ..parsers.base import MethodParser
+from ..parsers.base import MethodParser, TokenParser
 from ..parsers.chats import *
 
-from ..utils import handle_user_id, handle_chat_id
+from ..utils import get_current_user, handle_user_id, handle_chat_id
 
 
 class ChatResource(Resource):
     @staticmethod
     def get(chat_id):
         session = db_session.create_session()
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
         chat = handle_chat_id(chat_id, session)
+        if chat not in current_user.chats:
+            abort(401, message='You have no access to this Chat')
         return jsonify({'chat': chat.to_dict()})
 
     @staticmethod
@@ -43,18 +46,28 @@ class ChatResource(Resource):
 
     def post(self, chat_id):
         session = db_session.create_session()
-        handle_chat_id(chat_id, session)
+        chat = handle_chat_id(chat_id, session)
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
+        if chat not in current_user.chats:
+            abort(401, message='You have no access to this Chat')
         method_parser = MethodParser()
         method = method_parser.parse_args()['method']
         if method is not None:
-            parser = eval(f'Chat{method[0].upper() + method[1:]}Parser()')
-            args = eval(f'parser.parse_args()')
-            return eval(f'self.{method}(chat_id, **args)')
+            try:
+                parser = eval(f'Chat{method[0].upper() + method[1:]}Parser()')
+            except NameError:
+                abort(404, message='Unknown method was provided')
+            else:
+                args = parser.parse_args()
+                return eval(f'self.{method}(chat_id, **args)')
 
     @staticmethod
     def put(chat_id):
         session = db_session.create_session()
         chat = handle_chat_id(chat_id, session)
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
+        if chat not in current_user.chats:
+            abort(401, message='You have no access to this Chat')
         parser = ChatAddParser()
         args = parser.parse_args()
         if args['users'] is not None:
@@ -69,6 +82,9 @@ class ChatResource(Resource):
     def delete(chat_id):
         session = db_session.create_session()
         chat = handle_chat_id(chat_id, session)
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
+        if chat not in current_user.chats:
+            abort(401, message='You have no access to this Chat')
         session.delete(chat)
         session.commit()
         return jsonify({'message': 'OK'})
