+++ data/api/resources/user_resource.py
@@ -1,24 +1,30 @@
 from flask import jsonify
+from flask_jwt_extended import create_access_token
 from flask_restful import Resource, abort
 
 from ... import db_session
 from ...models.users import User
-# from ..parsers.base import MethodParser
 from ..parsers.users import *
+from ..parsers.base import TokenParser
 
-from ..utils import handle_user_id
+from ..utils import handle_user_id, get_current_user
 
 
 class UserResource(Resource):
     @staticmethod
     def get(user_id):
         session = db_session.create_session()
+        get_current_user(TokenParser().parse_args()['token'], session)
         user = handle_user_id(user_id, session)
         return jsonify({'user': user.to_dict()})
 
-    def put(self, user_id):
+    @staticmethod
+    def put(user_id):
         session = db_session.create_session()
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
         user = handle_user_id(user_id, session)
+        if current_user.id != user.id:
+            abort(401, message='You have no access to edit this User')
         parser = UserPutParser()
         args = parser.parse_args()
         if args.get('password') is not None:
@@ -32,7 +38,10 @@ class UserResource(Resource):
     @staticmethod
     def delete(user_id):
         session = db_session.create_session()
+        current_user = get_current_user(TokenParser().parse_args()['token'], session)
         user = handle_user_id(user_id, session)
+        if current_user.id != user.id:
+            abort(401, message='You have no access to delete this User')
         session.delete(user)
         session.commit()
         return jsonify({'message': 'OK'})
@@ -57,4 +66,4 @@ class UserPublicListResource(Resource):
         user.set_password(password)
         session.add(user)
         session.commit()
-        return jsonify({'message': 'OK'})
+        return jsonify({'token': create_access_token(identity=user.id)})
